Rename module, use positional args for Segment, rename value -> opacity in Opacity
